gui: split generic function of WaitingOverlay to a '_OverlayMixIn' class
