+++ src/silx/gui/widgets/WaitingOverlay.py
@@ -1,17 +1,17 @@
 from silx.gui.widgets.WaitingPushButton import WaitingPushButton
 from silx.gui import qt
-from silx.gui.qt import inspect as qt_inspect
-from silx.gui.plot import PlotWidget
+from .OverlayMixIn import _OverlayMixIn
 
 
-class WaitingOverlay(qt.QWidget):
+class WaitingOverlay(_OverlayMixIn, qt.QWidget):
     """Widget overlaying another widget with a processing wheel icon.
 
     :param parent: widget on top of which to display the "processing/waiting wheel"
     """
 
     def __init__(self, parent: qt.QWidget) -> None:
-        super().__init__(parent)
+        qt.QWidget.__init__(self, parent)
+        _OverlayMixIn.__init__(self, parent)
         self.setContentsMargins(0, 0, 0, 0)
 
         self._waitingButton = WaitingPushButton(self)
@@ -31,36 +31,6 @@ class WaitingOverlay(qt.QWidget):
         self._waitingButton.setText(text)
         self._resize()
 
-    def _listenedWidget(self, parent: qt.QWidget) -> qt.QWidget:
-        """Returns widget to register event filter to according to parent"""
-        if isinstance(parent, PlotWidget):
-            return parent.getWidgetHandle()
-        return parent
-
-    def _backendChanged(self):
-        self._listenedWidget(self.parent()).installEventFilter(self)
-        self._resizeLater()
-
-    def _registerParent(self, parent: qt.QWidget | None):
-        if parent is None:
-            return
-        self._listenedWidget(parent).installEventFilter(self)
-        if isinstance(parent, PlotWidget):
-            parent.sigBackendChanged.connect(self._backendChanged)
-        self._resize()
-
-    def _unregisterParent(self, parent: qt.QWidget | None):
-        if parent is None:
-            return
-        if isinstance(parent, PlotWidget):
-            parent.sigBackendChanged.disconnect(self._backendChanged)
-        self._listenedWidget(parent).removeEventFilter(self)
-
-    def setParent(self, parent: qt.QWidget):
-        self._unregisterParent(self.parent())
-        super().setParent(parent)
-        self._registerParent(parent)
-
     def showEvent(self, event: qt.QShowEvent):
         super().showEvent(event)
         self._waitingButton.setVisible(True)
@@ -69,41 +39,9 @@ class WaitingOverlay(qt.QWidget):
         super().hideEvent(event)
         self._waitingButton.setVisible(False)
 
-    def _resize(self):
-        if not qt_inspect.isValid(self):
-            return  # For _resizeLater in case the widget has been deleted
-
-        parent = self.parent()
-        if parent is None:
-            return
-
-        size = self._waitingButton.sizeHint()
-        if isinstance(parent, PlotWidget):
-            offset = parent.getWidgetHandle().mapTo(parent, qt.QPoint(0, 0))
-            left, top, width, height = parent.getPlotBoundsInPixels()
-            rect = qt.QRect(
-                qt.QPoint(
-                    int(offset.x() + left + width / 2 - size.width() / 2),
-                    int(offset.y() + top + height / 2 - size.height() / 2),
-                ),
-                size,
-            )
-        else:
-            position = parent.size()
-            position = (position - size) / 2
-            rect = qt.QRect(qt.QPoint(position.width(), position.height()), size)
-        self.setGeometry(rect)
-        self.raise_()
-
-    def _resizeLater(self):
-        qt.QTimer.singleShot(0, self._resize)
-
-    def eventFilter(self, watched: qt.QWidget, event: qt.QEvent):
-        if event.type() == qt.QEvent.Resize:
-            self._resize()
-            self._resizeLater()  # Defer resize for the receiver to have handled it
-        return super().eventFilter(watched, event)
-
     # expose Waiting push button API
     def setIconSize(self, size):
         self._waitingButton.setIconSize(size)
+
+    def sizeHin(self):
+        return self._waitingButton.sizeHint()
